transcritor-audiencias/requirements.txt
deepgram-sdk ~=3.3.1
openai-whisper ==20231117
python-dotenv ~=1.0.1
